fix parse error errors

+++ trunk-recorder/systems/p25_parser.cc
@@ -973,8 +973,8 @@ std::vector<TrunkMessage> P25Parser::parse_message(gr::message::sptr msg, System
 
   // # nac is always 1st two bytes
   // ac = (ord(s[0]) << 8) + ord(s[1])
-  uint8_t s0 = (int)s[0];
-  uint8_t s1 = (int)s[1];
+  uint16_t s0 = (int)s[0];
+  uint16_t s1 = (int)s[1];
   int shift = s0 << 8;
   long nac = shift + s1;
